requirements.txt
google-genai==1.62.0
chainlit==2.9.6
openpyxl==3.1.5
python-dotenv==1.1.0
